Update README.md
Team Aquaman Project Proposal
 
Title: Movie Recommendation System
 
Project Description:
Finding movies to watch is a struggle for people especially with how there is such an overwhelming amount of movies today. The movie recommendation system is a desktop application designed to get the preferences of a user having trouble to decide on what movies to watch. The user will be able to login to their account wherein they will be prompted to the menu to their current movie playlists/bookmarks or if they want to get a movie recommendation. By utilizing well-known websites such as rotten tomatoes as an example, the program will have an updated data set of movies with all their genres, subgenres, and ratings. 


 
Key-Features:
	1	Recommendation system
	2	Login system
	3	Bookmark/playlist for movies chosen